Refactored ecs infrastructure to use component ptr instead of component handles, and use new objects pool class

+++ src/ecs/TypeAliases.hpp
@@ -11,10 +11,10 @@ namespace detail
 template <class T> class MemoryPool;
 }
 
-using EntityId = uint32_t;
+using EntityId = int32_t;
 const std::size_t MaxComponentTypesCount = 128U;
 using ComponentMaskType = std::bitset<MaxComponentTypesCount>;
 using EntityHandleIndex = uint32_t;
-using ComponentTypeId = uint8_t;
+using ComponentTypeId = int32_t;
 
 }
